Add sleep functionality and refactor raw lock implementations
